rename files to be more logical
